Backend/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
billiard==3.6.4.0
bleach==6.0.0
blinker==1.4
cachelib==0.9.0
cairocffi==1.5.0
CairoSVG==2.7.0
celery==5.2.7
cffi==1.15.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.7.0
defusedxml==0.7.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.6.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==4.1.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
Pillow==9.4.0
pluggy==1.0.0
prompt-toolkit==3.0.38
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
pyphen==0.14.0
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.5.1
six==1.16.0
SQLAlchemy==1.4.37
tinycss2==1.2.1
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.6
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
